test: split video sort tests into a dedicated module
